Split two more pieces into modules (github, decisions)
